+++ Kuroba/app/src/main/java/com/github/adamantcheese/chan/ui/captcha/v2/nojs/CaptchaV2NoJsLayout.java
@@ -193,11 +193,11 @@ public class CaptchaV2NoJsLayout
 
             captchaImagesGrid.setAdapter(adapter);
             int columnsCount = captchaV2NoJsInfo.captchaType.columnCount;
-            int imageSize = Math.min(getWidth(), getHeight() - dp(104));
+            float imageSize = Math.min(getWidth(), getHeight() - dp(104));
             //40 + 64dp from layout xml; width for left-right full span, height minus for top-bottom full span inc buttons and titlebar
             ViewGroup.LayoutParams layoutParams = captchaImagesGrid.getLayoutParams();
-            layoutParams.height = imageSize;
-            layoutParams.width = imageSize;
+            layoutParams.height = (int) imageSize;
+            layoutParams.width = (int) imageSize;
             captchaImagesGrid.setLayoutParams(layoutParams);
 
             imageSize /= columnsCount;
