+++ Kuroba/app/src/main/java/com/github/adamantcheese/chan/ui/layout/DrawerWidthAdjustingLayout.java
@@ -46,9 +46,9 @@ public class DrawerWidthAdjustingLayout
 
         View drawer = findViewById(R.id.drawer);
 
-        int width = Math.min(widthSize - dp(getContext(), 56), dp(getContext(), 56) * 6);
+        float width = Math.min(widthSize - dp(getContext(), 56), dp(getContext(), 56) * 6);
         if (drawer.getLayoutParams().width != width) {
-            drawer.getLayoutParams().width = width;
+            drawer.getLayoutParams().width = (int) width;
         }
 
         super.onMeasure(widthMeasureSpec, heightMeasureSpec);
